Commit de Ronan Le Boulch : - Correcting Windows warnings
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/mptk/trunk@1987 0f77665a-23fb-0310-bc23-e91fd8efc3e1

+++ src/libmptk/convolution.cpp
@@ -1835,7 +1835,7 @@ void MP_Convolution_FFT_c::compute_IP( MP_Real_t* input, unsigned long int input
     numFramesNew = 0;
 
     firstFrameSample = nextFirstFrameSample;
-    nextFirstFrameSample = p - pNextSlice;    
+    nextFirstFrameSample = (unsigned long int)(p - pNextSlice);    
 
     while ((p < pNextSlice + anywaveTable->filterLen)&&(p < (pInputEnd))) {
       numFramesNew ++;
